requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
google-api-python-client==1.8.0
PyYAML==5.3.1
